Actually delete StandaloneBagVerifier
