+++ bag_verifier/src/main/scala/uk/ac/wellcome/platform/archive/bagverifier/services/BagVerifier.scala
@@ -12,24 +12,10 @@ import uk.ac.wellcome.platform.archive.common.bagit.services.BagReader
 import uk.ac.wellcome.platform.archive.common.ingests.models.IngestID
 import uk.ac.wellcome.platform.archive.common.storage.models._
 import uk.ac.wellcome.storage.listing.Listing
-import uk.ac.wellcome.storage.s3.{S3ObjectLocation, S3ObjectLocationPrefix}
+import uk.ac.wellcome.storage.s3.S3ObjectLocationPrefix
 import uk.ac.wellcome.storage.{Location, Prefix}
 
 import scala.util.Try
-
-trait StandaloneBagVerifier
-  extends BagVerifier[
-      StandaloneBagVerifyContext[S3ObjectLocationPrefix],
-      S3ObjectLocation,
-      S3ObjectLocationPrefix
-    ] {
-  override def verifyReplicatedBag(
-    root: StandaloneBagVerifyContext[S3ObjectLocationPrefix],
-    space: StorageSpace,
-    externalIdentifier: ExternalIdentifier,
-    bag: Bag
-  ): Either[BagVerifierError, Unit] = Right(())
-}
 
 trait ReplicatedBagVerifier[
   ReplicaBagLocation <: Location,
