+++ bag_verifier/src/test/scala/uk/ac/wellcome/platform/archive/bagverifier/services/BagVerifierTestCases.scala
@@ -20,7 +20,6 @@ import uk.ac.wellcome.platform.archive.common.bagit.services.{
   BagReader,
   BagUnavailable
 }
-import uk.ac.wellcome.platform.archive.common.fixtures.s3.S3BagBuilder
 import uk.ac.wellcome.platform.archive.common.fixtures.{
   BagBuilder,
   PayloadEntry
@@ -42,28 +41,6 @@ import uk.ac.wellcome.storage.store.TypedStore
 import uk.ac.wellcome.storage.store.fixtures.NamespaceFixtures
 import uk.ac.wellcome.storage.store.s3.S3TypedStore
 import uk.ac.wellcome.storage.{Location, Prefix}
-
-trait StandaloneBagVerifierTestCases[
-  BagLocation <: Location,
-  BagPrefix <: Prefix[
-    BagLocation
-  ],
-  Namespace
-] extends BagVerifierTestCases[
-      StandaloneBagVerifier,
-      StandaloneBagVerifyContext[BagPrefix],
-      BagLocation,
-      BagPrefix,
-      Namespace
-    ] {
-  override def withBagContext[R](
-    srcBagRoot: S3ObjectLocationPrefix,
-    replicaBagRoot: BagPrefix
-  )(testWith: TestWith[StandaloneBagVerifyContext[BagPrefix], R]): R =
-    testWith(
-      StandaloneBagVerifyContext(replicaBagRoot)
-    )
-}
 
 trait BagVerifierTestCases[Verifier <: BagVerifier[
   BagContext,
@@ -707,83 +684,4 @@ trait BagVerifierTestCases[Verifier <: BagVerifier[
 
       assertion(failedResult, summary)
     }
-}
-
-trait ReplicatedBagVerifierTestCases[
-  ReplicaBagLocation <: Location,
-  ReplicaBagPrefix <: Prefix[ReplicaBagLocation],
-  ReplicaNamespace
-] extends BagVerifierTestCases[
-      ReplicatedBagVerifier[
-        ReplicaBagLocation,
-        ReplicaBagPrefix
-      ],
-      ReplicatedBagVerifyContext[ReplicaBagPrefix],
-      ReplicaBagLocation,
-      ReplicaBagPrefix,
-      ReplicaNamespace
-    ] {
-  override def withBagContext[R](srcBagRoot: S3ObjectLocationPrefix, replicaBagRoot: ReplicaBagPrefix)(testWith: TestWith[ReplicatedBagVerifyContext[ReplicaBagPrefix], R]): R =
-    testWith(ReplicatedBagVerifyContext(srcBagRoot,replicaBagRoot))
-
-  it("fails a bag if it doesn't match original tag manifest") {
-    val space = createStorageSpace
-    val externalIdentifier = createExternalIdentifier
-    withTypedStore { implicit typedStore =>
-      withBag(space, externalIdentifier) { case (srcBucket, _, srcBagRoot, replicaBagRoot) =>
-
-        // Scramble the contents of the original tag manifest
-        putStream(srcBagRoot.asLocation("tagmanifest-sha256.txt"))
-
-        val ingestStep =
-          withVerifier(srcBucket) {
-            _.verify(
-              ingestId = createIngestID,
-              bagContext = ReplicatedBagVerifyContext(
-                srcRoot = srcBagRoot,
-                replicaRoot = replicaBagRoot
-              ),
-              space = space,
-              externalIdentifier = externalIdentifier
-            )
-          }
-
-        val result = ingestStep.success.get
-
-        result shouldBe a[IngestFailed[_]]
-        result.summary shouldBe a[VerificationIncompleteSummary]
-
-        result.maybeUserFacingMessage shouldNot be(defined)
-      }
-    }
-  }
-
-  it("fails a bag if it cannot read the original bag") {
-    val space = createStorageSpace
-    val externalIdentifier = createExternalIdentifier
-    withTypedStore { implicit typedStore =>
-      withBag(space, externalIdentifier) {
-        case (srcBucket, _, srcBagRoot, replicaBagRoot) =>
-          val ingestStep =
-            withVerifier(srcBucket) {
-              _.verify(
-                ingestId = createIngestID,
-                bagContext = ReplicatedBagVerifyContext(
-                  srcRoot = srcBagRoot,
-                  replicaRoot = replicaBagRoot
-                ),
-                space = space,
-                externalIdentifier = externalIdentifier
-              )
-            }
-
-          val result = ingestStep.success.get
-
-          result shouldBe a[IngestFailed[_]]
-          result.summary shouldBe a[VerificationIncompleteSummary]
-
-          result.maybeUserFacingMessage shouldNot be(defined)
-      }
-    }
-  }
 }
