+++ bag_verifier/src/test/scala/uk/ac/wellcome/platform/archive/bagverifier/services/s3/S3BagVerifierTest.scala
@@ -85,18 +85,19 @@ class S3ReplicatedBagVerifierTest
 }
 
 class S3StandaloneBagVerifierTest
-    extends StandaloneBagVerifierTestCases[
-      S3ObjectLocation,
-      S3ObjectLocationPrefix,
+    extends BagVerifierTestCases[
+      S3StandaloneBagVerifier,
+      StandaloneBagVerifyContext[S3ObjectLocationPrefix],
+      S3ObjectLocation, S3ObjectLocationPrefix,
       Bucket
     ]
     with S3BagVerifierTests[
-      StandaloneBagVerifier,
+      S3StandaloneBagVerifier,
       StandaloneBagVerifyContext[S3ObjectLocationPrefix]
     ] {
   override def withVerifier[R](primaryBucket: Bucket)(
     testWith: TestWith[
-      StandaloneBagVerifier,
+      S3StandaloneBagVerifier,
       R
     ]
   )(
@@ -107,4 +108,6 @@ class S3StandaloneBagVerifierTest
     )
 
   override val bagBuilder: BagBuilder[S3ObjectLocation, S3ObjectLocationPrefix, Bucket] = new S3BagBuilder {}
+
+  override def withBagContext[R](srcBagRoot: S3ObjectLocationPrefix, replicaBagRoot: S3ObjectLocationPrefix)(testWith: TestWith[StandaloneBagVerifyContext[S3ObjectLocationPrefix], R]): R = testWith(StandaloneBagVerifyContext(replicaBagRoot))
 }
